app/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
dnspython==2.7.0
dotenv==0.9.9
et_xmlfile==2.0.0
Faker==37.1.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
numpy==2.2.4
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pluggy==1.5.0
psycopg2-binary==2.9.10
pydantic==2.11.1
pydantic_core==2.33.0
pymongo==4.11.3
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
six==1.17.0
sniffio==1.3.1
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
